+++ changelog.md
@@ -4,3 +4,4 @@
 ## Prerelease version
 - created main game file
 - created a map
+- functional rooms